scripts/requirements.txt
Pillow>=10.0.0
matplotlib>=3.5.0
numpy>=1.21.0
noise>=1.2.2